Another fix for bug in [UIScrollView scrollViewToVisible:animated:]
We were having issues in which attempting to tap a button nested in
another view inside a UIScrollView caused erratic scrolling behavior.

The logic can now handle when views are nested inside a UIScrollView,
and takes its contentSize into account.

Note: this would fail the previous test added for this bug, because
the previous feature did not set the contentSize for the UIScrollView.